Extract interface and Remove log.print
