+++ commands/env.go
@@ -244,7 +244,7 @@ func detectShell() (string, error) {
 	if shell == "" {
 		// check for windows env and not bash (i.e. msysgit, etc)
 		if runtime.GOOS == "windows" {
-			log.Printf("On Windows, please specify either 'cmd' or 'powershell' with the --shell flag.\n\n")
+			log.Infof("On Windows, please specify either 'cmd' or 'powershell' with the --shell flag.\n\n")
 		}
 
 		return "", ErrUnknownShell
