+++ libmachine/log/interface.go
@@ -1,0 +1,20 @@
+package log
+
+type Logger interface {
+	Debug(...interface{})
+	Debugf(string, ...interface{})
+
+	Info(...interface{})
+	Infof(string, ...interface{})
+
+	Warn(...interface{})
+	Warnf(string, ...interface{})
+
+	Error(...interface{})
+	Errorf(string, ...interface{})
+
+	Fatal(...interface{})
+	Fatalf(string, ...interface{})
+
+	WithFields(Fields) Logger
+}
